renamed ooi and some files
